requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Faker==4.1.0
idna==2.9
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
